CurrentDate slice is now horizontalScroll
