+++ src/App.js
@@ -8,7 +8,7 @@ import globals from "./globals";
 import mocks from "./mocks";
 import "./App.css";
 import { useDispatch } from "react-redux";
-import { scroll, setDate } from './currentDateSlice';
+import { scroll, setCurrentDate } from './horizontalScrollSlice';
 
 function App(props) {
   const hotel = mocks.hotel;
@@ -24,7 +24,7 @@ function App(props) {
       date: new Date(event.target.value),
       tiles: tiles
     });
-    dispatch(setDate({ newDateString: event.target.value }));
+    dispatch(setCurrentDate({ newDateString: event.target.value }));
   }
 
   function onTileMove(event) {
